fix unit test to get 100% coverage, timezone test becomes client test as the builder and the parser are already tested
